+++ src/handlers/defect_dojo_handler.py
@@ -59,7 +59,8 @@ class DefectDojoHandler(AbstractHandler):
             Endpoint.INTEGRATIONS_DEFECT_DOJO_ID_ACTIVATION: {
                 HTTPMethod.PUT: self.put_activation,
                 HTTPMethod.DELETE: self.delete_activation,
-                HTTPMethod.GET: self.get_activation
+                HTTPMethod.GET: self.get_activation,
+                HTTPMethod.PATCH: self.patch_activation
             }
         }
 
@@ -216,3 +217,59 @@ class DefectDojoHandler(AbstractHandler):
         # first because they are all equal
         meta = DefectDojoParentMeta.from_dict(parents[0].meta.as_dict())
         return build_response(self.get_dto(parents, meta))
+
+    @validate_kwargs
+    def patch_activation(self, event: DefectDojoActivationPutModel, id: str,
+                         _pe: ProcessedEvent):
+
+        item = self._dds.get_nullable(id)
+        if not item or event.customer and item.customer != event.customer:
+            raise ResponseFactory(HTTPStatus.NOT_FOUND).message(
+                self._dds.not_found_message(id)
+            ).exc()
+
+        if event.tenant_names:
+            it = iter_tenants_by_names(
+                tenant_service=self._ps.tenant_service,
+                customer=event.customer_id,
+                names=event.tenant_names,
+                attributes_to_get=(Tenant.name, )
+            )
+            tenants = {tenant.name for tenant in it}
+            if missing := event.tenant_names - tenants:
+                raise ResponseFactory(HTTPStatus.NOT_FOUND).message(
+                    f'Active tenant(s) {", ".join(missing)} not found'
+                ).exc()
+
+        parents_list = list(self.get_all_activations(item.id, event.customer))
+
+        meta = DefectDojoParentMeta(
+            scan_type=event.scan_type,
+            product_type=event.product_type,
+            product=event.product,
+            engagement=event.engagement,
+            test=event.test,
+            send_after_job=event.send_after_job,
+            attachment=event.attachment
+        )
+
+        to_update = build_parents(
+            payload=ResolveParentsPayload(
+                parents=parents_list,
+                tenant_names=event.tenant_names,
+                exclude_tenants=event.exclude_tenants,
+                clouds=event.clouds,
+                all_tenants=event.all_tenants
+            ),
+            parent_service=self._ps,
+            application_id=id,
+            customer_id=event.customer,
+            type_=ParentType.CUSTODIAN_SIEM_DEFECT_DOJO,
+            created_by=_pe['cognito_user_id'],
+            meta=meta.dict()
+        )
+
+        for parent in to_update:
+            self._ps.update(parent)
+        return build_response(content=self.get_dto(to_update, meta))
+