+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `category` field to `OPERATIONAL RULES` report to distinguish rules with identical fingerprints
 - Added fingerprint to the `Rule` model and `SRERules` collection for rule deduplication, optimizing API requests to cloud providers
 - Added the ability to provide receivers to the endpoint `/report/department/`
+- Added new CLI `sre integrations dojo update` command
 
 ### Fixed
 - Fixed invalid `next_token` handling